modified aop.xml settings; dataservices moved to separate packages
