+++ backend-core/src/main/java/moe/ofs/backend/chatcmdnew/BasicChatCommandBootstrap.java
@@ -7,7 +7,7 @@ import moe.ofs.backend.domain.dcs.poll.PlayerInfo;
 import moe.ofs.backend.function.triggermessage.model.TriggerMessage;
 import moe.ofs.backend.function.triggermessage.services.NetMessageService;
 import moe.ofs.backend.function.triggermessage.services.TriggerMessageService;
-import moe.ofs.backend.dataservice.PlayerInfoService;
+import moe.ofs.backend.dataservice.player.PlayerInfoService;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.PostConstruct;
