+++ backend-core/src/main/java/moe/ofs/backend/discipline/controllers/PlayerInfoListController.java
@@ -1,7 +1,7 @@
 package moe.ofs.backend.discipline.controllers;
 
 import moe.ofs.backend.domain.dcs.poll.PlayerInfo;
-import moe.ofs.backend.dataservice.PlayerInfoService;
+import moe.ofs.backend.dataservice.player.PlayerInfoService;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
