+++ backend-core/src/main/java/moe/ofs/backend/function/newslotcontrol/services/SlotManageServiceImpl.java
@@ -4,9 +4,9 @@ import lombok.extern.slf4j.Slf4j;
 import moe.ofs.backend.domain.dcs.poll.ExportObject;
 import moe.ofs.backend.domain.dcs.poll.PlayerInfo;
 import moe.ofs.backend.object.FlyableUnit;
-import moe.ofs.backend.dataservice.ExportObjectService;
-import moe.ofs.backend.dataservice.FlyableUnitService;
-import moe.ofs.backend.dataservice.PlayerInfoService;
+import moe.ofs.backend.dataservice.exportobject.ExportObjectService;
+import moe.ofs.backend.dataservice.slotunit.FlyableUnitService;
+import moe.ofs.backend.dataservice.player.PlayerInfoService;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.PostConstruct;
