+++ backend-core/src/main/java/moe/ofs/backend/function/triggermessage/model/MessageQueue.java
@@ -2,7 +2,7 @@ package moe.ofs.backend.function.triggermessage.model;
 
 import moe.ofs.backend.domain.dcs.poll.ExportObject;
 import moe.ofs.backend.function.triggermessage.services.TriggerMessageService;
-import moe.ofs.backend.dataservice.FlyableUnitService;
+import moe.ofs.backend.dataservice.slotunit.FlyableUnitService;
 import org.springframework.lang.NonNull;
 
 import java.util.ArrayDeque;
