+++ backend-core/src/main/java/moe/ofs/backend/function/triggermessage/services/impl/NotificationMessageServiceImpl.java
@@ -6,8 +6,8 @@ import moe.ofs.backend.function.triggermessage.model.TriggerMessage;
 import moe.ofs.backend.function.triggermessage.services.NetMessageService;
 import moe.ofs.backend.function.triggermessage.services.NotificationMessageService;
 import moe.ofs.backend.function.triggermessage.services.TriggerMessageService;
-import moe.ofs.backend.dataservice.ExportObjectService;
-import moe.ofs.backend.dataservice.FlyableUnitService;
+import moe.ofs.backend.dataservice.exportobject.ExportObjectService;
+import moe.ofs.backend.dataservice.slotunit.FlyableUnitService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
