+++ backend-core/src/main/java/moe/ofs/backend/function/triggermessage/services/impl/TriggerMessageServiceImpl.java
@@ -7,8 +7,8 @@ import moe.ofs.backend.function.triggermessage.PlayerNotActiveException;
 import moe.ofs.backend.function.triggermessage.model.TriggerMessage;
 import moe.ofs.backend.function.triggermessage.services.TriggerMessageService;
 import moe.ofs.backend.object.FlyableUnit;
-import moe.ofs.backend.dataservice.FlyableUnitService;
-import moe.ofs.backend.dataservice.PlayerInfoService;
+import moe.ofs.backend.dataservice.slotunit.FlyableUnitService;
+import moe.ofs.backend.dataservice.player.PlayerInfoService;
 import moe.ofs.backend.connector.util.LuaScripts;
 import moe.ofs.backend.connector.lua.LuaQueryEnv;
 import org.springframework.stereotype.Service;
