+++ backend-core/src/main/java/moe/ofs/backend/services/jpa/ExportObjectBulkJpaService.java
@@ -6,7 +6,7 @@
 //import moe.ofs.backend.handlers.ExportUnitDespawnObservable;
 //import moe.ofs.backend.handlers.ExportUnitSpawnObservable;
 //import moe.ofs.backend.common.ExportObjectRepository;
-//import moe.ofs.backend.dataservice.ExportObjectService;
+//import moe.ofs.backend.dataservice.exportobject.ExportObjectService;
 //import org.springframework.stereotype.Service;
 //
 //import java.util.HashSet;
