+++ backend-data/src/main/java/moe/ofs/backend/dataservice/aspect/GraveyardCollectAspectConfiguration.java
@@ -1,7 +1,6 @@
 package moe.ofs.backend.dataservice.aspect;
 
-import moe.ofs.backend.dataservice.GraveyardService;
-import org.aspectj.lang.Aspects;
+import moe.ofs.backend.dataservice.graveyard.GraveyardService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
