+++ backend-data/src/main/java/moe/ofs/backend/dataservice/aware/PlayerInfoServiceAware.java
@@ -1,6 +1,6 @@
 package moe.ofs.backend.dataservice.aware;
 
-import moe.ofs.backend.dataservice.PlayerInfoService;
+import moe.ofs.backend.dataservice.player.PlayerInfoService;
 import org.springframework.beans.factory.Aware;
 
 public interface PlayerInfoServiceAware extends Aware {
