+++ backend-data/src/main/java/moe/ofs/backend/dataservice/exportobject/ExportObjectService.java
@@ -1,4 +1,4 @@
-package moe.ofs.backend.dataservice;
+package moe.ofs.backend.dataservice.exportobject;
 
 import moe.ofs.backend.common.CrudService;
 import moe.ofs.backend.common.MissionPersistenceRepository;
