+++ backend-data/src/main/java/moe/ofs/backend/dataservice/slotunit/FlyableUnitMapService.java
@@ -1,4 +1,4 @@
-package moe.ofs.backend.dataservice.map;
+package moe.ofs.backend.dataservice.slotunit;
 
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
@@ -10,7 +10,6 @@ import moe.ofs.backend.connector.ConnectionManager;
 import moe.ofs.backend.connector.util.LuaScripts;
 import moe.ofs.backend.connector.response.JsonRpcResponse;
 import moe.ofs.backend.connector.request.server.ServerExecRequest;
-import moe.ofs.backend.dataservice.FlyableUnitService;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
