+++ pom.xml
@@ -140,43 +140,6 @@
                 </plugin>
 
                 <plugin>
-                    <groupId>org.apache.maven.plugins</groupId>
-                    <artifactId>maven-antrun-plugin</artifactId>
-                    <version>${maven-antrun-plugin.version}</version>
-                    <executions>
-                        <execution>
-                            <id>unwovenClassesFolder</id>
-                            <phase>generate-resources</phase>
-                            <configuration>
-                                <target>
-                                    <delete dir="${project.build.directory}/unwoven-classes"/>
-                                    <mkdir dir="${project.build.directory}/unwoven-classes"/>
-                                </target>
-                            </configuration>
-                            <goals>
-                                <goal>run</goal>
-                            </goals>
-                        </execution>
-                    </executions>
-                </plugin>
-
-                <plugin>
-                    <groupId>org.apache.maven.plugins</groupId>
-                    <artifactId>maven-compiler-plugin</artifactId>
-                    <version>${maven-compiler-plugin.version}</version>
-                    <executions>
-                        <execution>
-                            <id>default-compile</id>
-                            <configuration>
-                                <compilerArguments>
-                                    <d>${project.build.directory}/unwoven-classes</d>
-                                </compilerArguments>
-                            </configuration>
-                        </execution>
-                    </executions>
-                </plugin>
-
-                <plugin>
                     <groupId>com.github.m50d</groupId>
                     <artifactId>aspectj-maven-plugin</artifactId>
                     <version>1.11.1</version>
@@ -216,7 +179,7 @@
                         <Xlint>ignore</Xlint>
                         <encoding>${project.build.sourceEncoding}</encoding>
                         <forceAjcCompile>true</forceAjcCompile>
-<!--                        <sources/>&lt;!&ndash; this is important!&ndash;&gt;-->
+                        <sources/><!-- this is important!-->
                     </configuration>
                 </plugin>
